+++ lib/src/commands/can/can_push.dart
@@ -8,7 +8,7 @@ import 'package:gg_one_state/gg_one_state.dart';
 import 'package:gg_args/gg_args.dart';
 import 'package:gg_log/gg_log.dart';
 import 'package:gg_one/src/commands/can/can_commit.dart';
-import 'package:gg_one/src/tools/checks.dart';
+import 'package:gg_one_checks/gg_one_checks.dart';
 
 /// Are the last changes ready for »git push«?
 class CanPush extends CommandCluster {