+++ lib/src/commands/do/do_publish.dart
@@ -16,7 +16,6 @@ import 'package:gg_log/gg_log.dart';
 import 'package:gg_merge/gg_merge.dart' as gg_merge;
 import 'package:gg_lang/gg_lang.dart';
 import 'package:gg_one/src/commands/can/can_publish.dart';
-import 'package:gg_one/src/commands/check/no_pubspec_overrides.dart';
 import 'package:gg_one/src/commands/did/did_commit.dart';
 import 'package:gg_one/src/commands/do/do_configure_publish.dart';
 import 'package:gg_one/src/commands/do/do_push.dart';
@@ -31,6 +30,7 @@ import 'package:gg_publish/gg_publish.dart';
 import 'package:gg_version/gg_version.dart';
 import 'package:path/path.dart';
 import 'package:pub_semver/pub_semver.dart';
+import 'package:gg_one_checks/gg_one_checks.dart';
 
 /// Publishes the current directory.
 ///